Lib

Description

This is an experimental playground for how to maintain documentation for a REST endpoint and associated datatypes, and hosting the associated documentation for both.

Documentation

runServer :: IO ()

data CreateTodoRequest

This is the internal representation of the components of the todo create request.

Constructors

CreateTodoRequest

Fields

ctrTitle :: Text

Title of the todo. Quick, snappy description of the task.

ctrDescription :: Text

More in-depth description of the work to be done.

data CreateTodoResponse

Internal representation of the response of the create todo request. If successful, this will return the todo ID for the user to refer to when they want to refer to or update the todo later.

Constructors

CreateTodoResponse

Fields

ctrTid :: Int

ID of the todo that was created.

data FailedToCreateTodo

If the create request fails, return the request that failed. Don't expose the internal error, as that's a security vulnerability. But, we should add some facility to communicate broadly what went wrong, especially if the failure was due to invalid inputs.

Constructors

FailedToCreateTodo

Fields

createReq :: CreateTodoRequest

The originating create request type.

data TodosStruct

This struct is used to represent the result of a read request. As such, it has the title, description, and completed fields you'd expect, as well as the todo ID.

Constructors

TodosStruct

Fields

todoTitle :: Text

Title of the todo. Quick, snappy description of the task.

todoDescription :: Text

More in-depth description of the work to be done.

todoCompleted :: Bool

Has the task been completed? True for yes.

tsTid :: Int

What ID did the database assign to this task?

getHomeR :: Handler Html

The handler for the base route.

getReadAllTodosR :: Handler Value

postRestCreateR :: Handler Value

Endpoint to create a TODO. Consumes a JSON body that matches up with the CreateTodoRequest type. Returns a CreateTodoResponse with the Todo ID, or a FailedToCreateTodo message on failure.