Types

Description

The types that are used in Lib are split out here, to make both files a bit less noisy and easier to work with.

Documentation

data CreateTodoRequest

This is the internal representation of the components of the todo create request.

Constructors

CreateTodoRequest

Fields

ctrTitle :: Text

Title of the todo. Quick, snappy description of the task.

ctrDescription :: Text

More in-depth description of the work to be done.

data CreateTodoResponse

Internal representation of the response of the create todo request. If successful, this will return the todo ID for the user to refer to when they want to refer to or update the todo later.

Constructors

CreateTodoResponse

Fields

ctrTid :: Int

ID of the todo that was created.

data FailedToCreateTodo

If the create request fails, return the request that failed. Don't expose the internal error, as that's a security vulnerability. But, we should add some facility to communicate broadly what went wrong, especially if the failure was due to invalid inputs.

Constructors

FailedToCreateTodo

Fields

createReq :: CreateTodoRequest

The originating create request type.

data TodosStruct

This struct is used to represent the result of a read request. As such, it has the title, description, and completed fields you'd expect, as well as the todo ID.

Constructors

TodosStruct

Fields

todoTitle :: Text

Title of the todo. Quick, snappy description of the task.

todoDescription :: Text

More in-depth description of the work to be done.

todoCompleted :: Bool

Has the task been completed? True for yes.

tsTid :: Int

What ID did the database assign to this task?

data TodoNotFound

This struct is returned as a JSON object if we try to read, update, or delete a todo ID that cannot be found.

Constructors

TodoNotFound

Fields

tnfTid :: Int

The missing Todo ID

data MultipleTodosFound

We return this type if we run afoul of the invariant that each todo has a unique ID. This really only exists to handle a circumstance that the types allow, but should be impossible.

Constructors

MultipleTodosFound

Fields

mtfTid :: Int

The TID that somehow was found more than once.

data UpdateTodoRequest

An update request. If a field is Nothing, then it will be unchanged in the request. Note that this allows for a "no-op" update, but having the types constrain that "at least one field shall be populated" would be ... tricky.

Constructors

UpdateTodoRequest

Fields

utrTitle :: Maybe Text

A possible change to a title.

utrDescription :: Maybe Text

A possible change to the description.

utrFinished :: Maybe Bool

A possible change to the finished state.

data UpdateTodoResponse

The response returned -- as a JSON object -- on successfully updating a todo entry.

Constructors

UpdateTodoResponse

Fields

utrTid :: Int

Update was successful, return TID update was performed on

data DeleteTodoResponse

The response returned -- as a JSON object -- on successfully deleting a todo entry. Hopefully one we've finished, rather than abandoned.

Constructors

DeleteTodoResponse

Fields

dtrTid :: Int

Update was successful, return TID update was performed on